Move table definiitons into tables module
