+++ packages/grid/apps/domain/src/main/core/database/tables/bin_obj.py
@@ -3,8 +3,8 @@ from syft import deserialize
 from syft import serialize
 
 # grid relative
-from . import BaseModel
-from . import db
+from .. import BaseModel
+from .. import db
 
 
 class BinObject(BaseModel):
