+++ packages/grid/apps/domain/src/main/core/database/tables/dataset_group.py
@@ -1,6 +1,6 @@
 # grid relative
-from . import BaseModel
-from . import db
+from .. import BaseModel
+from .. import db
 
 
 class DatasetGroup(BaseModel):
