+++ packages/grid/apps/domain/src/main/core/database/utils.py
@@ -1,8 +1,8 @@
 # grid relative
 from . import db
-from .groups.groups import Group
-from .groups.usergroup import UserGroup
-from .roles.roles import Role
+from .tables.group import Group
+from .tables.user_group import UserGroup
+from .tables.role import Role
 
 
 def model_to_json(model):
