+++ packages/grid/apps/domain/src/main/core/dataset_ops.py
@@ -20,13 +20,13 @@ import torch as th
 
 # grid relative
 from .database import db
-from .database.bin_obj import BinObject
-from .database.bin_obj import ObjectMetadata
-from .database.json_obj import JsonObject
-from .database.metadata import get_metadata
-from .database.bin_obj_dataset import BinObjDataset
-from .database.dataset import Dataset
-from .database.dataset_group import DatasetGroup
+from .database.tables.bin_obj import BinObject
+from .database.tables.bin_obj import ObjectMetadata
+from .database.tables.json_obj import JsonObject
+from .database.tables.metadata import get_metadata
+from .database.tables.bin_obj_dataset import BinObjDataset
+from .database.tables.dataset import Dataset
+from .database.tables.dataset_group import DatasetGroup
 from .database.store_disk import DiskObjectStore
 from .database.utils import model_to_json
