+++ packages/grid/apps/domain/src/main/core/manager/association_request_manager.py
@@ -5,8 +5,8 @@ from typing import List
 from typing import Union
 
 # grid relative
-from ..database.association import Association
-from ..database.association_request import AssociationRequest
+from ..database.tables.association import Association
+from ..database.tables.association_request import AssociationRequest
 from ..exceptions import AssociationRequestError
 from .database_manager import DatabaseManager
