+++ packages/grid/apps/domain/src/main/core/manager/group_manager.py
@@ -2,15 +2,9 @@
 from typing import List
 
 # grid relative
-from ..database.groups.groups import Group
-from ..database.groups.usergroup import UserGroup
-from ..exceptions import AuthorizationError
+from ..database.tables.group import Group
+from ..database.tables.user_group import UserGroup
 from ..exceptions import GroupNotFoundError
-from ..exceptions import InvalidCredentialsError
-from ..exceptions import MissingRequestKeyError
-from ..exceptions import PyGridError
-from ..exceptions import RoleNotFoundError
-from ..exceptions import UserNotFoundError
 from .database_manager import DatabaseManager
